+++ module/product/domain/models.go
@@ -1,6 +1,10 @@
 package domain
 
-import "time"
+import (
+	"time"
+
+	"github.com/capstone-kelompok-7/backend-disappear/module/review/domain"
+)
 
 type Product struct {
 	ID            int             `gorm:"column:id;type:int;primaryKey" json:"id"`
@@ -15,7 +19,7 @@ type Product struct {
 	UpdatedAt     time.Time       `gorm:"column:updated_at;type:TIMESTAMP" json:"updated_at"`
 	DeletedAt     *time.Time      `gorm:"column:deleted_at;type:TIMESTAMP;index" json:"deleted_at"`
 	ProductPhotos []ProductPhotos `gorm:"foreignKey:ProductID" json:"product_photos"`
-	ProductReview []Review        `gorm:"foreignKey:ProductID;references:ID" json:"review"`
+	ProductReview []domain.Review `gorm:"foreignKey:ProductID;references:ID" json:"review"`
 	Categories    []Category      `gorm:"many2many:product_categories;" json:"categories"`
 }
 
@@ -36,14 +40,4 @@ type ProductPhotos struct {
 	CreatedAt time.Time  `gorm:"column:created_at;type:TIMESTAMP" json:"created_at"`
 	UpdatedAt time.Time  `gorm:"column:updated_at;type:TIMESTAMP" json:"updated_at"`
 	DeletedAt *time.Time `gorm:"column:deleted_at;type:TIMESTAMP;index" json:"deleted_at"`
-}
-
-type Review struct {
-	ID          int        `gorm:"column:id;type:int;primaryKey" json:"id"`
-	ProductID   int        `gorm:"column:product_id;type:int" json:"product_id"`
-	Description string     `gorm:"column:description;type:text" json:"description"`
-	Date        time.Time  `gorm:"column:date;type:date" json:"date"`
-	CreatedAt   time.Time  `gorm:"column:created_at;type:TIMESTAMP" json:"created_at"`
-	UpdatedAt   time.Time  `gorm:"column:updated_at;type:TIMESTAMP" json:"updated_at"`
-	DeletedAt   *time.Time `gorm:"column:deleted_at;type:TIMESTAMP;index" json:"deleted_at"`
 }
